Refactor orderCode generation in PaymentService
The code change modifies the way the `orderCode` is generated in the `CreatePaymentLinkAsync` method of the `PaymentService` class. Previously, the `orderCode` was generated by parsing the current timestamp formatted as "ffffff" (microseconds) into an integer. The new code changes this to parse the `orderId` from the `CreatePaymentLinkRequest` object as a long integer instead.