Add endpoint to fetch confirmed orders with pagination
A new endpoint has been added to the `OrderController` class to get a list of orders with the status `CONFIRMED`. This endpoint is accessible via an HTTP GET request to the `confirmed` route and supports pagination with `pageNumber` and `pageSize` parameters.

In the `IOrderRepository` interface, a new method `GetListOrderConfirmed` has been added to fetch orders with the status `CONFIRMED`. The `OrderRepository` class has been updated to implement this method, retrieving orders with the status `CONFIRMED` from the database.

In the `IOrderService` interface, a new method `GetListOrderConfirmed` has been added to fetch a paginated list of orders with the status `CONFIRMED`. The `OrderService` class has been updated to implement this method, using the repository method to fetch the orders and then returning a paginated list.